moved domain tests
