+++ tests/CookieTest.php
@@ -51,32 +51,6 @@ final class CookieTest extends TestCase
         self::assertEmpty($newCookieFromOptions->getValue());
     }
 
-    public function testNormalizeDomain(): void
-    {
-        $domain1 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('www.google.de');
-        $domain2 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('www.google');
-        $domain3 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('google.de');
-        $domain4 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('domain-@test.@de');
-        $domain5 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('xn--fsqu00a.xn--0zwm56d');
-        $domain6 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('..');
-        $domain7 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('.');
-        $domain8 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('-_-');
-        $domain9 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('---');
-        $domain10 = marcoschwepp\Cookie\CookieUtility::normalizeDomain('');
-
-        self::assertSame($domain1, '.www.google.de');
-        self::assertSame($domain2, '.www.google');
-        self::assertSame($domain3, '.google.de');
-        self::assertSame($domain5, '.xn--fsqu00a.xn--0zwm56d');
-
-        self::assertNull($domain4);
-        self::assertNull($domain6);
-        self::assertNull($domain7);
-        self::assertNull($domain8);
-        self::assertNull($domain9);
-        self::assertNull($domain10);
-    }
-
     public function testGettersAndSetters(): void
     {
         $options = [
